LayoutTests/http/tests/site-isolation/open-panel-in-cross-origin-iframe-expected.txt

OPEN FILE PANEL
Tests that choosing a file in a cross-origin <input type=file> routes the open-panel reply to the iframe's WebContent process when site isolation is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS chosenFile is "foo.txt"
PASS successfullyParsed is true

TEST COMPLETE
